refactor(trainer): split validation + optimizers out, extract SFT-switch clutter, dedupe mixture weights
trainer.py (425 -> 309 lines) now holds only what a training run IS — the
loop and its data pipeline. Moved, not rewritten:
- validation.py: _val_ce_sums + ValidationProbe (+ its constants); the probe
  now takes data_root explicitly instead of reading trainer's global.
- optimizers.py: weight_decay_mask, the base optimizer_chain, and
  create_sft_optimizer — both chains built by one _make_chain so the SFT
  variant differs by exactly the LR schedule, as the comment always claimed.

In-loop cleanups:
- The SFT phase switch reads top-to-bottom: early-exit halt, banner, flag,
  then the OOM-critical host-staged swap (kept inline — a helper's del
  cannot release the loop's local reference, which is what the #30 fix
  depends on), then monitor.reset_for_new_phase(...) instead of the trainer
  groping five monitor attributes.
- LossMonitor.push documents its return (plateau signal) and is_new_best
  side channel; reset_for_new_phase keeps the checkpoint-serialized
  attribute set stable.
- Mixture weights deduped: pretrain uses schedules.CURRICULUM_START_WEIGHTS,
  SFT uses new schedules.SFT_MIX_WEIGHTS in both the mixer and the log line
  (which previously hardcoded a copy of the list as a string).

No numeric change: golden-run, determinism, and the full suite pass; the
exact swap path (chain -> steps -> plateau -> host-staged rebuild -> steps)
exercised end-to-end on CPU. CLAUDE.md repo map updated.
